finport-ma-plot/requirements.txt
finsim>=1.0.5
boto3>=1.17.0
plotnine>=0.10.0
openpyxl>=3.0.0
python-dotenv
